Add files via upload
Chimera states occur when there is the coexistence of spatially homogeneous and inhomogeneous sectors in spatially extended dynamical systems, like coupled
map lattices. To characterize chimeras, however, is a nontrivial problem since it is difficult to distinguish between homogeneous and inhomogeneous states. A
useful tool for this task is machine learning, in particular deep learning techniques like reservoir computing and multilayer perceptrons. In this work we use
these quantifiers in order to identify chimera states in logistic map lattices with non-local coupling. We compare our results from machine learning techniques
with more convencional characterizations, such as Lyapunov exponents and a local order parameter.